+++ src/hover-provider.js
@@ -4,16 +4,43 @@ import { formatNumber } from './formatter.js';
 
 class HoverProvider {
     provideHover(document, position) {
-        const range = document.getWordRangeAtPosition(position, /(0x)?\d+(\.\d+)?/),
-            word = range && document.getText(range);
+        // Regular numbers
+        {
+            const range = document.getWordRangeAtPosition(position, /-?(0x)?\d+(\.\d+)?/),
+                word = range && document.getText(range);
 
-        if (!word) return;
+            if (word) {
+                const isFloat = word.includes('.'),
+                    number = isFloat ? parseFloat(word) : parseInt(word);
 
-        const isFloat = word.includes('.'),
-            number = isFloat ? parseFloat(word) : parseInt(word);
+                return new vscode.Hover(formatNumber(number));
+            }
+        }
 
-        return new vscode.Hover(formatNumber(number));
+        // Expressions
+        {
+            const range = document.getWordRangeAtPosition(position, /[\d *+\-\/^().]+/),
+                word = range && document.getText(range);
+
+            // Is it a somewhat valid expression?
+            if (word && !word.match(/^[^\d]+$|[-+*\/().]{2,}/m)) {
+                let result;
+
+                // Try to evaluate the expression
+                try {
+                    result = evalMath(word);
+
+                    result = result && ('= ' + formatNumber(result));
+                } catch (e) { }
+
+                return new vscode.Hover(result || 'Invalid expression');
+            }
+        }
     }
+}
+
+function evalMath(str) {
+    return Function(`'use strict'; return (${str})`)()
 }
 
 export function registerHoverProvider(context) {